Specify more information on installing xformers, give basic example
xformers has 3 different versions. If a user isn't using their cpu for inferencing, then they will have to uninstall xformers and reinstall the correct version after visit xformers README.md. Further, the namespace to which the function `enable_xformers_memory_efficient_attention()` was never specified. For all a new user could know its under the torch or diffusers namespace, when in actuality it exists on a pipeline.

This edit is simply to save headaches and time, as one who has fallen into this trap.